Integrate enhanced pulse dashboard and drop _new prototypes
