--- _new/mt5_bridge.py
@@ -1,311 +1,0 @@
-"""
-MT5 Bridge - Real Trade History Integration for Pulse UI
-Connects to MT5 account history and provides real behavioral analytics
-"""
-
-import MetaTrader5 as mt5
-import pandas as pd
-import numpy as np
-from datetime import datetime, timedelta
-from typing import Dict, List, Optional, Tuple
-import json
-import yaml
-
-class MT5Bridge:
-    """
-    Bridge between MT5 trading platform and Pulse system.
-    Pulls real trade history, analyzes behavioral patterns, and feeds the dashboard.
-    """
-
-    def __init__(self, account: int = None, password: str = None, server: str = None):
-        """Initialize MT5 connection."""
-        self.account = account
-        self.connected = False
-        self.trade_cache = {}
-        self.behavioral_patterns = {}
-
-    def connect(self) -> bool:
-        """Establish connection to MT5."""
-        if not mt5.initialize():
-            print("MT5 initialization failed")
-            return False
-
-        if self.account:
-            authorized = mt5.login(self.account, password=self.password, server=self.server)
-            if not authorized:
-                print(f"Failed to connect to account {self.account}")
-                mt5.shutdown()
-                return False
-
-        self.connected = True
-        return True
-
-    def get_real_trade_history(self, days_back: int = 30) -> pd.DataFrame:
-        """
-        Fetch real trade history from MT5.
-
-        Returns:
-            DataFrame with actual trades, not mock data
-        """
-        if not self.connected:
-            raise ConnectionError("Not connected to MT5")
-
-        # Get deals from history
-        from_date = datetime.now() - timedelta(days=days_back)
-        deals = mt5.history_deals_get(from_date, datetime.now())
-
-        if deals is None or len(deals) == 0:
-            return pd.DataFrame()
-
-        # Convert to DataFrame
-        df = pd.DataFrame(list(deals), columns=deals[0]._asdict().keys())
-
-        # Process and enrich with behavioral data
-        df = self._enrich_with_behavioral_data(df)
-
-        return df
-
-    def _enrich_with_behavioral_data(self, df: pd.DataFrame) -> pd.DataFrame:
-        """
-        Enrich trade history with behavioral analysis.
-        This is where the magic happens - turning raw trades into behavioral insights.
-        """
-        df['time'] = pd.to_datetime(df['time'], unit='s')
-        df = df.sort_values('time')
-
-        # Calculate consecutive wins/losses
-        df['is_win'] = df['profit'] > 0
-        df['consecutive_wins'] = df.groupby((df['is_win'] != df['is_win'].shift()).cumsum())['is_win'].cumsum()
-        df['consecutive_losses'] = df.groupby((~df['is_win'] != ~df['is_win'].shift()).cumsum()['is_win'].apply(lambda x: (~x).cumsum())
-
-        # Detect revenge trading (quick re-entry after loss)
-        df['time_since_last'] = df['time'].diff()
-        df['revenge_trade'] = (df['profit'].shift() < 0) & (df['time_since_last'] < pd.Timedelta(minutes=15))
-
-        # Detect overconfidence (increased position after wins)
-        df['position_change'] = df['volume'].pct_change()
-        df['overconfidence'] = (df['consecutive_wins'] >= 3) & (df['position_change'] > 0.5)
-
-        # Time-based patterns
-        df['hour'] = df['time'].dt.hour
-        df['is_optimal_time'] = df['hour'].between(8, 20)  # London/NY sessions
-        df['fatigue_trade'] = df.groupby(df['time'].dt.date).cumcount() > 5  # More than 5 trades/day
-
-        # FOMO detection (chasing after missing moves)
-        df['entry_speed'] = df['time_since_last'].dt.total_seconds()
-        df['fomo_trade'] = df['entry_speed'] < 60  # Entry within 1 minute of last trade
-
-        # Calculate running P&L for the day
-        df['daily_pnl'] = df.groupby(df['time'].dt.date)['profit'].cumsum()
-        df['daily_drawdown'] = df.groupby(df['time'].dt.date)['daily_pnl'].cummin()
-
-        # Risk score based on behavioral flags
-        df['behavioral_risk_score'] = (
-            df['revenge_trade'].astype(int) * 3 +
-            df['overconfidence'].astype(int) * 2 +
-            df['fatigue_trade'].astype(int) * 2 +
-            df['fomo_trade'].astype(int) * 1 +
-            (~df['is_optimal_time']).astype(int) * 1
-        )
-
-        return df
-
-    def get_behavioral_report(self) -> Dict:
-        """
-        Generate comprehensive behavioral analysis report.
-        This is what makes the dashboard powerful - real patterns, not theory.
-        """
-        df = self.get_real_trade_history()
-
-        if df.empty:
-            return {"error": "No trade history available"}
-
-        report = {
-            "performance_metrics": {
-                "total_trades": len(df),
-                "win_rate": (df['is_win'].sum() / len(df) * 100),
-                "avg_win": df[df['is_win']]['profit'].mean(),
-                "avg_loss": df[~df['is_win']]['profit'].mean(),
-                "profit_factor": abs(df[df['is_win']]['profit'].sum() / df[~df['is_win']]['profit'].sum()),
-                "max_consecutive_wins": df['consecutive_wins'].max(),
-                "max_consecutive_losses": df['consecutive_losses'].max(),
-                "total_pnl": df['profit'].sum()
-            },
-            "behavioral_patterns": {
-                "revenge_trades": {
-                    "count": df['revenge_trade'].sum(),
-                    "success_rate": df[df['revenge_trade']]['is_win'].mean() * 100 if df['revenge_trade'].any() else 0,
-                    "avg_loss": df[df['revenge_trade'] & ~df['is_win']]['profit'].mean()
-                },
-                "overconfidence_trades": {
-                    "count": df['overconfidence'].sum(),
-                    "success_rate": df[df['overconfidence']]['is_win'].mean() * 100 if df['overconfidence'].any() else 0,
-                    "avg_result": df[df['overconfidence']]['profit'].mean()
-                },
-                "fatigue_trades": {
-                    "count": df['fatigue_trade'].sum(),
-                    "success_rate": df[df['fatigue_trade']]['is_win'].mean() * 100 if df['fatigue_trade'].any() else 0,
-                    "typical_time": df[df['fatigue_trade']]['hour'].mode().values[0] if df['fatigue_trade'].any() else None
-                },
-                "fomo_trades": {
-                    "count": df['fomo_trade'].sum(),
-                    "avg_loss": df[df['fomo_trade'] & ~df['is_win']]['profit'].mean()
-                }
-            },
-            "time_analysis": {
-                "best_hours": df.groupby('hour')['profit'].sum().nlargest(3).to_dict(),
-                "worst_hours": df.groupby('hour')['profit'].sum().nsmallest(3).to_dict(),
-                "trades_by_hour": df.groupby('hour').size().to_dict()
-            },
-            "risk_violations": {
-                "overtrade_days": len(df.groupby(df['time'].dt.date).size()[df.groupby(df['time'].dt.date).size() > 5]),
-                "max_daily_trades": df.groupby(df['time'].dt.date).size().max(),
-                "drawdown_breaches": len(df[df['daily_drawdown'] < -300])  # $300 daily limit
-            }
-        }
-
-        return report
-
-    def get_confluence_validation(self) -> pd.DataFrame:
-        """
-        Validate historical confluence scores against actual outcomes.
-        This proves whether the scoring system actually works.
-        """
-        df = self.get_real_trade_history()
-
-        if df.empty:
-            return pd.DataFrame()
-
-        # Simulate confluence scores (in production, these would be stored)
-        # This is where you'd pull actual scores from your signal journal
-        df['confluence_score'] = np.random.randint(40, 95, size=len(df))
-
-        # Bin confluence scores
-        df['confluence_band'] = pd.cut(df['confluence_score'], 
-                                       bins=[0, 50, 60, 70, 80, 90, 100],
-                                       labels=['<50', '50-60', '60-70', '70-80', '80-90', '90+'])
-
-        # Calculate win rate by confluence band
-        validation = df.groupby('confluence_band').agg({
-            'is_win': ['count', 'mean'],
-            'profit': ['sum', 'mean']
-        }).round(2)
-
-        validation.columns = ['trades', 'win_rate', 'total_pnl', 'avg_pnl']
-        validation['win_rate'] = validation['win_rate'] * 100
-
-        return validation
-
-    def sync_to_pulse_journal(self, journal_path: str = "signal_journal.json"):
-        """
-        Sync MT5 trade history with Pulse signal journal.
-        This creates the accountability loop.
-        """
-        df = self.get_real_trade_history()
-
-        if df.empty:
-            return
-
-        # Load existing journal
-        try:
-            with open(journal_path, 'r') as f:
-                journal = json.load(f)
-        except FileNotFoundError:
-            journal = []
-
-        # Convert trades to journal entries
-        for _, trade in df.iterrows():
-            entry = {
-                "timestamp": trade['time'].isoformat(),
-                "type": "mt5_trade",
-                "data": {
-                    "ticket": int(trade['ticket']),
-                    "symbol": trade['symbol'] if 'symbol' in trade else "UNKNOWN",
-                    "volume": float(trade['volume']),
-                    "profit": float(trade['profit']),
-                    "is_win": bool(trade['is_win']),
-                    "behavioral_flags": {
-                        "revenge_trade": bool(trade['revenge_trade']),
-                        "overconfidence": bool(trade['overconfidence']),
-                        "fatigue_trade": bool(trade['fatigue_trade']),
-                        "fomo_trade": bool(trade['fomo_trade'])
-                    },
-                    "behavioral_risk_score": int(trade['behavioral_risk_score']),
-                    "consecutive_wins": int(trade['consecutive_wins']),
-                    "consecutive_losses": int(trade['consecutive_losses'])
-                }
-            }
-
-            # Check if entry already exists
-            if not any(e.get('data', {}).get('ticket') == entry['data']['ticket'] for e in journal):
-                journal.append(entry)
-
-        # Save updated journal
-        with open(journal_path, 'w') as f:
-            json.dump(journal, f, indent=2)
-
-        return len(journal)
-
-    def generate_weekly_review(self) -> Dict:
-        """
-        Generate weekly performance and behavioral review.
-        This is the reflection engine - turning data into wisdom.
-        """
-        df = self.get_real_trade_history(days_back=7)
-
-        if df.empty:
-            return {"error": "No trades this week"}
-
-        review = {
-            "week_ending": datetime.now().strftime("%Y-%m-%d"),
-            "summary": {
-                "total_trades": len(df),
-                "total_pnl": df['profit'].sum(),
-                "win_rate": (df['is_win'].mean() * 100),
-                "best_day": df.groupby(df['time'].dt.date)['profit'].sum().idxmax().strftime("%Y-%m-%d"),
-                "worst_day": df.groupby(df['time'].dt.date)['profit'].sum().idxmin().strftime("%Y-%m-%d")
-            },
-            "behavioral_insights": {
-                "revenge_trading_incidents": df['revenge_trade'].sum(),
-                "overconfidence_incidents": df['overconfidence'].sum(),
-                "fatigue_incidents": df['fatigue_trade'].sum(),
-                "average_risk_score": df['behavioral_risk_score'].mean()
-            },
-            "recommendations": []
-        }
-
-        # Generate personalized recommendations
-        if df['revenge_trade'].sum() > 2:
-            review["recommendations"].append("Implement mandatory 30-min cooldown after losses")
-
-        if df['overconfidence'].sum() > 3:
-            review["recommendations"].append("Reduce position size by 50% after 3 consecutive wins")
-
-        if df['fatigue_trade'].sum() > 5:
-            review["recommendations"].append("Limit daily trades to 5 maximum")
-
-        if df[~df['is_optimal_time']].sum() > len(df) * 0.3:
-            review["recommendations"].append("Focus trading on London/NY session hours only")
-
-        return review
-
-    def disconnect(self):
-        """Disconnect from MT5."""
-        if self.connected:
-            mt5.shutdown()
-            self.connected = False
-
-
-# Example usage and integration
-if __name__ == "__main__":
-    # Initialize bridge (replace with your MT5 credentials)
-    bridge = MT5Bridge()
-
-    # For demo purposes, create sample integration
-    print("MT5 Bridge initialized")
-    print("This bridge provides:")
-    print("- Real trade history with behavioral analysis")
-    print("- Pattern detection (revenge, overconfidence, fatigue, FOMO)")
-    print("- Confluence score validation against outcomes")
-    print("- Weekly behavioral reviews")
-    print("- Direct sync with Pulse signal journal")
