--- _new/pulse_config.yaml
@@ -1,75 +1,0 @@
-# Zanalytics Pulse Configuration
-# Version: 11.5.1
-
-system:
-  name: "Zanalytics Pulse"
-  version: "11.5.1"
-  environment: "production"
-
-redis:
-  host: "localhost"
-  port: 6379
-  db: 0
-  password: null
-
-confluence_weights:
-  smc: 0.4          # Smart Money Concepts
-  wyckoff: 0.3      # Wyckoff Analysis
-  technical: 0.3    # Technical Indicators
-
-risk_limits:
-  daily_loss_limit: 500           # USD
-  daily_loss_percent: 0.03        # 3% of account
-  max_trades_per_day: 5
-  max_position_size: 0.02         # 2% risk per trade
-  cooling_period_minutes: 15
-  revenge_trade_window: 30        # minutes
-  max_consecutive_losses: 3
-
-behavioral:
-  revenge_trade_threshold: 3
-  overconfidence_threshold: 4
-  fatigue_hour: 22
-  optimal_hours: [9, 10, 11, 14, 15, 16]
-
-position_breakdown:
-  enabled: true
-  default_splits: 5               # Split position into 5 parts
-  min_confluence_per_entry: 60    # Minimum score for each entry
-  scale_in_increment: 5           # Confluence points between entries
-
-discord:
-  enabled: true
-  webhook_url: "${DISCORD_WEBHOOK_URL}"
-  alert_types:
-    - high_confluence
-    - risk_warning
-    - behavioral_alert
-    - daily_summary
-
-ui:
-  refresh_rate: 2                 # seconds
-  theme: "dark"
-  show_behavioral_state: true
-  show_position_breakdown: true
-  show_journal_prompts: true
-
-journal:
-  auto_prompt: true
-  prompt_on_loss: true
-  prompt_on_win_streak: true
-  weekly_review: true
-  storage_path: "/data/journal"
-
-monitoring:
-  metrics_port: 9090
-  health_check_interval: 30      # seconds
-  alert_on_error: true
-
-api_endpoints:
-  pulse_frame: "/api/pulse/frame"
-  pulse_health: "/api/pulse/health"
-  pulse_score: "/api/pulse/score"
-  pulse_risk: "/api/pulse/risk"
-  pulse_journal: "/api/pulse/journal"
-  signals_top: "/api/signals/top"
