--- _new/risk_enforcer.py
@@ -1,310 +1,0 @@
-"""
-RiskEnforcer - Behavioral Protection System
-Enforces trading discipline and risk management rules
-"""
-
-from datetime import datetime, timedelta
-from typing import Dict, List, Optional
-import json
-import logging
-
-logger = logging.getLogger(__name__)
-
-class RiskEnforcer:
-    """
-    Enforces risk management rules and behavioral safeguards.
-    Acts as the "seatbelt" for trading decisions.
-    """
-
-    def __init__(self, config: Dict = None):
-        # Risk limits (configurable)
-        self.limits = config.get('limits', {
-            'daily_loss_limit': 500,        # USD
-            'daily_loss_percent': 0.03,     # 3% (buffer before 5% prop limit)
-            'max_trades_per_day': 5,
-            'max_position_size': 0.02,      # 2% max risk per trade
-            'cooling_period_minutes': 15,
-            'revenge_trade_window': 30,     # minutes
-            'max_consecutive_losses': 3
-        }) if config else self._default_limits()
-
-        # Current state
-        self.daily_stats = {
-            'date': datetime.now().date(),
-            'trades_count': 0,
-            'total_pnl': 0,
-            'losses_count': 0,
-            'consecutive_losses': 0,
-            'last_loss_time': None,
-            'cooling_until': None,
-            'violations': []
-        }
-
-        # Behavioral tracking
-        self.behavioral_flags = {
-            'revenge_trading': False,
-            'overconfidence': False,
-            'fatigue': False,
-            'tilt': False
-        }
-
-    def _default_limits(self) -> Dict:
-        """Default risk limits"""
-        return {
-            'daily_loss_limit': 500,
-            'daily_loss_percent': 0.03,
-            'max_trades_per_day': 5,
-            'max_position_size': 0.02,
-            'cooling_period_minutes': 15,
-            'revenge_trade_window': 30,
-            'max_consecutive_losses': 3
-        }
-
-    def allow(self, signal: Dict) -> Tuple[bool, List[str], Dict]:
-        """
-        Main enforcement method - checks if a trade should be allowed.
-
-        Args:
-            signal: Trade signal with score, size, etc.
-
-        Returns:
-            Tuple of (allowed, warnings, details)
-        """
-        allowed = True
-        warnings = []
-        details = {}
-
-        # Check daily loss limit
-        loss_check = self._check_daily_loss()
-        if not loss_check['allowed']:
-            allowed = False
-            warnings.append(loss_check['reason'])
-
-        # Check trade count
-        count_check = self._check_trade_count()
-        if not count_check['allowed']:
-            allowed = False
-            warnings.append(count_check['reason'])
-        elif count_check.get('warning'):
-            warnings.append(count_check['warning'])
-
-        # Check cooling period
-        cooling_check = self._check_cooling_period()
-        if not cooling_check['allowed']:
-            allowed = False
-            warnings.append(cooling_check['reason'])
-
-        # Check position size
-        size_check = self._check_position_size(signal.get('size', 0))
-        if not size_check['allowed']:
-            allowed = False
-            warnings.append(size_check['reason'])
-
-        # Check behavioral patterns
-        behavioral_check = self._check_behavioral_patterns()
-        if behavioral_check['warnings']:
-            warnings.extend(behavioral_check['warnings'])
-
-        # Check for revenge trading
-        revenge_check = self._check_revenge_trading()
-        if revenge_check['detected']:
-            warnings.append('⚠️ Revenge trading pattern detected')
-            self.behavioral_flags['revenge_trading'] = True
-
-        # Check for overconfidence
-        confidence_check = self._check_overconfidence()
-        if confidence_check['detected']:
-            warnings.append('⚠️ Overconfidence detected - consider smaller size')
-            self.behavioral_flags['overconfidence'] = True
-
-        # Compile details
-        details = {
-            'trades_remaining': self.limits['max_trades_per_day'] - self.daily_stats['trades_count'],
-            'loss_buffer': self.limits['daily_loss_limit'] - abs(self.daily_stats['total_pnl']),
-            'behavioral_score': self._calculate_behavioral_score(),
-            'risk_level': self._calculate_risk_level(),
-            'flags': self.behavioral_flags
-        }
-
-        return allowed, warnings, details
-
-    def _check_daily_loss(self) -> Dict:
-        """Check if daily loss limit is exceeded"""
-        if self.daily_stats['total_pnl'] <= -self.limits['daily_loss_limit']:
-            return {
-                'allowed': False,
-                'reason': '🔴 Daily loss limit reached. No more trades today.'
-            }
-
-        # Warning if close to limit
-        if self.daily_stats['total_pnl'] <= -self.limits['daily_loss_limit'] * 0.8:
-            return {
-                'allowed': True,
-                'warning': '⚠️ Approaching daily loss limit (80% reached)'
-            }
-
-        return {'allowed': True}
-
-    def _check_trade_count(self) -> Dict:
-        """Check if max trades per day exceeded"""
-        if self.daily_stats['trades_count'] >= self.limits['max_trades_per_day']:
-            return {
-                'allowed': False,
-                'reason': '🔴 Maximum trades per day reached.'
-            }
-
-        if self.daily_stats['trades_count'] == self.limits['max_trades_per_day'] - 1:
-            return {
-                'allowed': True,
-                'warning': '⚠️ This is your last trade for today'
-            }
-
-        return {'allowed': True}
-
-    def _check_cooling_period(self) -> Dict:
-        """Check if cooling period is active"""
-        if self.daily_stats['cooling_until']:
-            if datetime.now() < self.daily_stats['cooling_until']:
-                remaining = (self.daily_stats['cooling_until'] - datetime.now()).seconds // 60
-                return {
-                    'allowed': False,
-                    'reason': f'❄️ Cooling period active. {remaining} minutes remaining.'
-                }
-            else:
-                # Cooling period expired
-                self.daily_stats['cooling_until'] = None
-
-        return {'allowed': True}
-
-    def _check_position_size(self, size: float) -> Dict:
-        """Check if position size is within limits"""
-        if size > self.limits['max_position_size']:
-            return {
-                'allowed': False,
-                'reason': f'🔴 Position size {size:.1%} exceeds maximum {self.limits["max_position_size"]:.1%}'
-            }
-        return {'allowed': True}
-
-    def _check_behavioral_patterns(self) -> Dict:
-        """Check for problematic behavioral patterns"""
-        warnings = []
-
-        # Check time of day (fatigue)
-        current_hour = datetime.now().hour
-        if current_hour >= 22 or current_hour <= 6:
-            warnings.append('🌙 Late night trading - higher risk of errors')
-            self.behavioral_flags['fatigue'] = True
-
-        # Check consecutive losses
-        if self.daily_stats['consecutive_losses'] >= 2:
-            warnings.append(f'📉 {self.daily_stats["consecutive_losses"]} consecutive losses - consider a break')
-
-        # Check rapid trading
-        if self._is_rapid_trading():
-            warnings.append('⚡ Rapid trading detected - slow down')
-
-        return {'warnings': warnings}
-
-    def _check_revenge_trading(self) -> Dict:
-        """Detect revenge trading patterns"""
-        if not self.daily_stats['last_loss_time']:
-            return {'detected': False}
-
-        time_since_loss = datetime.now() - self.daily_stats['last_loss_time']
-
-        if time_since_loss.seconds < self.limits['revenge_trade_window'] * 60:
-            if self.daily_stats['consecutive_losses'] > 0:
-                return {
-                    'detected': True,
-                    'confidence': 0.8,
-                    'reason': 'Quick re-entry after loss'
-                }
-
-        return {'detected': False}
-
-    def _check_overconfidence(self) -> Dict:
-        """Detect overconfidence patterns"""
-        # Simple check: many trades in succession or after wins
-        if self.daily_stats['trades_count'] >= 4:
-            return {
-                'detected': True,
-                'confidence': 0.7,
-                'reason': 'High trade frequency'
-            }
-        return {'detected': False}
-
-    def _is_rapid_trading(self) -> bool:
-        """Check if trading too rapidly"""
-        # Would check timestamps of recent trades
-        # Simplified for now
-        return False
-
-    def _calculate_behavioral_score(self) -> int:
-        """Calculate overall behavioral score 0-100"""
-        score = 100
-
-        # Deduct for flags
-        if self.behavioral_flags['revenge_trading']:
-            score -= 30
-        if self.behavioral_flags['overconfidence']:
-            score -= 20
-        if self.behavioral_flags['fatigue']:
-            score -= 15
-        if self.behavioral_flags['tilt']:
-            score -= 25
-
-        # Deduct for violations
-        score -= len(self.daily_stats['violations']) * 10
-
-        return max(0, score)
-
-    def _calculate_risk_level(self) -> str:
-        """Calculate current risk level"""
-        score = self._calculate_behavioral_score()
-
-        if score >= 80:
-            return 'low'
-        elif score >= 60:
-            return 'medium'
-        elif score >= 40:
-            return 'high'
-        else:
-            return 'critical'
-
-    def update_trade_result(self, result: Dict):
-        """Update stats after a trade completes"""
-        self.daily_stats['trades_count'] += 1
-        self.daily_stats['total_pnl'] += result.get('pnl', 0)
-
-        if result.get('pnl', 0) < 0:
-            self.daily_stats['losses_count'] += 1
-            self.daily_stats['consecutive_losses'] += 1
-            self.daily_stats['last_loss_time'] = datetime.now()
-
-            # Trigger cooling period after significant loss
-            if abs(result.get('pnl', 0)) > 100:
-                self.daily_stats['cooling_until'] = (
-                    datetime.now() + timedelta(minutes=self.limits['cooling_period_minutes'])
-                )
-        else:
-            self.daily_stats['consecutive_losses'] = 0
-
-    def reset_daily_stats(self):
-        """Reset daily statistics"""
-        self.daily_stats = {
-            'date': datetime.now().date(),
-            'trades_count': 0,
-            'total_pnl': 0,
-            'losses_count': 0,
-            'consecutive_losses': 0,
-            'last_loss_time': None,
-            'cooling_until': None,
-            'violations': []
-        }
-
-        self.behavioral_flags = {
-            'revenge_trading': False,
-            'overconfidence': False,
-            'fatigue': False,
-            'tilt': False
-        }
